+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Automatically process choices with a `RUN_ALL` option into a list.
+- Add `with_overwrite` cli option.
 
 ## [1.0.12] - 2024-11-13
 ### Fixed